requirement.txt
argparse==1.2.1
Django==1.5
South==0.8.4
httplib2==0.8
oauth2==1.5.211
oauthlib==0.6.0
python-twitter==1.1
requests==2.0.1
requests-oauthlib==0.4.0
simplejson==3.3.1
